test: move gpu_query from manual to impls/mpich
The manual tests are not regularly tested. Modify the gpu_query to be
able to run by default and move into impls/mpich so it can be tested
regularly.
