+++ test/mpi/impls/mpich/Makefile.am
@@ -7,7 +7,7 @@ include $(top_srcdir)/Makefile_single.mtest
 
 EXTRA_DIST = testlist.in
 
-static_subdirs = mpi_t comm
+static_subdirs = mpi_t comm misc
 # For future tests - note that the IO and RMA directories are optional,
 # and need to be handled as shown in this comment
 #SUBDIRS      = $(static_subdirs) $(iodir) $(rmadir)
