+++ test/mpi/impls/mpich/misc/testlist
@@ -1,0 +1,1 @@
+gpu_query 1
