+++ test/mpi/impls/mpich/testlist.in
@@ -1,5 +1,6 @@
 mpi_t
 comm
+misc
 @threadsdir@
 # The IO and RMA directories are optional and need to be handled in
 # a special way.  Uncomment these as appropriate when a test is added.
